+++ src/components/navbars/topbar/api/pickApi.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react';
-import { ApiStatus, RootState} from '../../redux/types';
+import { ApiStatus, RootState} from '../../../../redux/types';
 import { useDispatch, useSelector } from 'react-redux';
 import swal from 'sweetalert';
-import { createTheme, ThemeProvider, ListItemButton, ListItemText, ListItemIcon, Collapse, ErrorIcon, ExpandLess, ExpandMore, CancelIcon, IconButton, DoNotDisturbOnIcon, CheckCircleIcon } from '../../muiImports' 
+import { createTheme, ThemeProvider, ListItemButton, ListItemText, ListItemIcon, Collapse, ErrorIcon, ExpandLess, ExpandMore, CancelIcon, IconButton, DoNotDisturbOnIcon, CheckCircleIcon } from '../../../../muiImports' 
 
 // import StreamTextDropdown from './streamtext/streamTextDropdown';
-import AzureDropdown from './azure/AzureDropdown';
+import AzureDropdown from './AzureDropdown';
 
 enum STATUS {
     "AVAILABLE",
